+++ src/main/java/com/gromed/demo/repository/GeneriqueRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GeneriqueRepository extends JpaRepository<Generique, Integer> {
+public interface GeneriqueRepository extends JpaRepository<Generique, Long> {
 }
