+++ src/main/java/com/gromed/demo/repository/MedicamentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicamentRepository extends JpaRepository<Medicament, Integer> {
+public interface MedicamentRepository extends JpaRepository<Medicament, Long> {
 }
